backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cloudinary==1.44.1
cryptography==45.0.5
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
frozenlist==1.7.0
google-auth==2.40.3
gunicorn==23.0.0
idna==3.10
mercadopago==2.3.0
multidict==6.6.3
oauthlib==3.3.1
packaging==25.0
propcache==0.3.2
psycopg==3.2.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
stream-chat==4.26.0
twilio==9.6.5
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.2.3
whitenoise==6.9.0
yarl==1.20.1
